Implement graphemes with generic
