+++ graphemes/fuzz_test.go
@@ -37,7 +37,7 @@ func FuzzValidShort(f *testing.F) {
 		valid1 := utf8.Valid(original)
 		tokens := graphemes.FromBytes(original)
 		for tokens.Next() {
-			all = append(all, tokens.Bytes())
+			all = append(all, tokens.Value())
 		}
 
 		roundtrip := make([]byte, 0, len(original))
@@ -77,7 +77,7 @@ func FuzzValidLong(f *testing.F) {
 		valid1 := utf8.Valid(original)
 		tokens := graphemes.FromBytes(original)
 		for tokens.Next() {
-			all = append(all, tokens.Bytes())
+			all = append(all, tokens.Value())
 		}
 
 		roundtrip := make([]byte, 0, len(original))
@@ -135,7 +135,7 @@ func FuzzInvalid(f *testing.F) {
 		valid1 := utf8.Valid(original)
 		tokens := graphemes.FromBytes(original)
 		for tokens.Next() {
-			all = append(all, tokens.Bytes())
+			all = append(all, tokens.Value())
 		}
 
 		roundtrip := make([]byte, 0, len(original))
