+++ graphemes/splitfunc.go
@@ -1,6 +1,10 @@
 package graphemes
 
-var trie = &graphemesTrie{}
+import (
+	"bufio"
+
+	"github.com/clipperhouse/uax29/v2/internal/iterators"
+)
 
 // is determines if lookup intersects propert(ies)
 func (lookup property) is(properties property) bool {
@@ -9,14 +13,20 @@ func (lookup property) is(properties property) bool {
 
 const _Ignore = _Extend
 
-// SplitFunc is a bufio.SplitFunc implementation of Unicode grapheme cluster
-// segmentation, for use with bufio.Scanner.
+// SplitFunc is a bufio.SplitFunc implementation of Unicode grapheme cluster segmentation, for use with bufio.Scanner.
 //
 // See https://unicode.org/reports/tr29/#Grapheme_Cluster_Boundaries.
-func SplitFunc(data []byte, atEOF bool) (advance int, token []byte, err error) {
+var SplitFunc bufio.SplitFunc = func(data []byte, atEOF bool) (advance int, token []byte, err error) {
+	return splitFunc(data, atEOF)
+}
+
+func splitFunc[T iterators.Stringish](data T, atEOF bool) (advance int, token T, err error) {
+	var empty T
 	if len(data) == 0 {
-		return 0, nil, nil
+		return 0, empty, nil
 	}
+
+	trie := &graphemesTrie[T]{}
 
 	// These vars are stateful across loop iterations
 	var pos int
@@ -31,7 +41,7 @@ func SplitFunc(data []byte, atEOF bool) (advance int, token []byte, err error) {
 	if w == 0 {
 		if !atEOF {
 			// Rune extends past current data, request more
-			return 0, nil, nil
+			return 0, empty, nil
 		}
 		pos = len(data)
 		return pos, data[:pos], nil
@@ -47,7 +57,7 @@ func SplitFunc(data []byte, atEOF bool) (advance int, token []byte, err error) {
 		if eot {
 			if !atEOF {
 				// Token extends past current data, request more
-				return 0, nil, nil
+				return 0, empty, nil
 			}
 
 			// https://unicode.org/reports/tr29/#GB2
@@ -78,7 +88,7 @@ func SplitFunc(data []byte, atEOF bool) (advance int, token []byte, err error) {
 				break
 			}
 			// Rune extends past current data, request more
-			return 0, nil, nil
+			return 0, empty, nil
 		}
 
 		// Optimization: no rule can possibly apply
