+++ graphemes/trie.go
@@ -1,5 +1,7 @@
 package graphemes
 
+import "github.com/clipperhouse/uax29/v2/internal/iterators"
+
 // generated by github.com/clipperhouse/uax29/v2
 // from https://www.unicode.org/Public/15.0.0/ucd/auxiliary/GraphemeBreakProperty.txt
 
@@ -25,7 +27,7 @@ const (
 // lookup returns the trie value for the first UTF-8 encoding in s and
 // the width in bytes of this encoding. The size will be 0 if s does not
 // hold enough bytes to complete the encoding. len(s) must be greater than 0.
-func (t *graphemesTrie) lookup(s []byte) (v property, sz int) {
+func (t *graphemesTrie[T]) lookup(s T) (v property, sz int) {
 	c0 := s[0]
 	switch {
 	case c0 < 0x80: // is ASCII
@@ -86,14 +88,14 @@ func (t *graphemesTrie) lookup(s []byte) (v property, sz int) {
 }
 
 // graphemesTrie. Total size: 29120 bytes (28.44 KiB). Checksum: 80ad0c5ab9375f7.
-type graphemesTrie struct{}
+type graphemesTrie[T iterators.Stringish] struct{}
 
 // func newGraphemesTrie(i int) *graphemesTrie {
 // 	return &graphemesTrie{}
 // }
 
 // lookupValue determines the type of block n and looks up the value for b.
-func (t *graphemesTrie) lookupValue(n uint32, b byte) property {
+func (t *graphemesTrie[T]) lookupValue(n uint32, b byte) property {
 	switch {
 	default:
 		return property(graphemesValues[n<<6+uint32(b)])
