+++ internal/iterators/iterator.go
