+++ words/joiners_test.go
@@ -1,6 +1,10 @@
 package words_test
 
-import "github.com/clipperhouse/uax29/v2/words"
+import (
+	"testing"
+
+	"github.com/clipperhouse/uax29/v2/words"
+)
 
 var joinersInput = []byte("Hello, 世界. Tell me about your super-cool .com. I'm .01% interested and 3/4 of a mile away. Email me at foo@example.biz. #winning")
 var joiners = &words.Joiners[[]byte]{
@@ -36,4 +40,67 @@ var joinersTests = []joinersTest{
 	{"#", true, false},
 	{"winning", true, false},
 	{"#winning", false, true},
+}
+
+func TestGenericIteratorWithJoiners(t *testing.T) {
+	t.Parallel()
+
+	// Test with []byte and joiners
+	text := []byte("Hello, 世界. Tell me about your super-cool .com. I'm .01% interested and 3/4 of a mile away. Email me at foo@example.biz. #winning")
+	iter := words.FromBytes(text)
+
+	joiners := &words.Joiners[[]byte]{
+		Middle:  []rune("@-/"),
+		Leading: []rune("#."),
+	}
+	iter.Joiners(joiners)
+
+	var results []string
+	for iter.Next() {
+		results = append(results, string(iter.Value()))
+	}
+
+	// Check that some specific joined tokens exist
+	expectedJoined := []string{"super-cool", ".com", ".01", "3/4", "foo@example.biz", "#winning"}
+	for _, expected := range expectedJoined {
+		found := false
+		for _, result := range results {
+			if result == expected {
+				found = true
+				break
+			}
+		}
+		if !found {
+			t.Errorf("Expected joined token %q not found in results", expected)
+		}
+	}
+
+	// Test with string and joiners
+	textStr := string(text)
+	iterStr := words.FromString(textStr)
+
+	joinersStr := &words.Joiners[string]{
+		Middle:  []rune("@-/"),
+		Leading: []rune("#."),
+	}
+	iterStr.Joiners(joinersStr)
+
+	var resultsStr []string
+	for iterStr.Next() {
+		resultsStr = append(resultsStr, iterStr.Value())
+	}
+
+	// Check that some specific joined tokens exist
+	for _, expected := range expectedJoined {
+		found := false
+		for _, result := range resultsStr {
+			if result == expected {
+				found = true
+				break
+			}
+		}
+		if !found {
+			t.Errorf("Expected joined token %q not found in string results", expected)
+		}
+	}
 }
